Refactor main.py and organize code into separate modules.
Suggested and byt the ChatGPT itselft 🤯:
- Move FastAPI routes from main.py to modules/routes.py
- Move Pydantic models to modules/models.py
- Move utility function to modules/utils.py
- Simplify main.py to only initialize and run FastAPI server
